Add files via upload
In This Project i have done data analysis on world bank debt. It's not that we humans only take debts to manage our necessities. A country may also take debt to manage its economy. For example, infrastructure spending is one costly ingredient required for a country's citizens to lead comfortable lives. The dataset contains information about the amount of debt (in USD).
In this Power BI dash Board we can see , 
What is the total amount of debt that is owed by the countries listed in the dataset.
Which country owns the maximum amount of debt and what does that amount look like.
What is the average amount of debt owed by countries across different debt indicators.
Number of distinct countries
Totaling the amount of debt owed by the countries.
Average amount of debt across indicators.